Accumulate lanyon_assets rather than overwriting
because with caching, we sometimes only get passed JS assets, which effectively removes CSS asset references

/cc @lakesare relevant to your interest. I had a similar issue that you were experiencing and chased it down to this. It may take a few changes in different asset types for `content/.lanyon/jekyll.lanyon_assets.yml` to feature both js and css files. After that, Jekyll will have an easier time inlining them.

And I was wrong. While it used to be that app.js was enough, the way it's set up now, we do need a css file. Changes are laid on top via js, but not from zero.

I think this also means, it will fix the flickering you experienced